+++ tests/test_prg_engine_runtime_surface_functions_tests.h
@@ -17,6 +17,7 @@ namespace copperfin::runtime_surface_tests
     void test_expression_runtime_surface_extensions();
     void test_curval_evaluates_on_disk_record();
     void test_curval_uses_verified_post_commit_session_image();
+    void test_lupdate_reads_table_header_date_and_designator_variants();
     void test_curval_verified_commit_overlay_does_not_survive_cursor_reopen();
     void test_getfldstate_tracks_buffered_mutation_state();
     void test_setfldstate_assigns_buffered_mutation_state();